fix: correct MCP schemas and error signaling (#46)
* fix: restore schema descriptions, align transaction enums, mark tool errors, tighten input validation

Supersedes #27; thanks to @jdmadison for identifying the cleared/flagColor schema mismatch.

- Re-attach property-level descriptions after the null-compatibility wrapper
  in the registry so they survive in the advertised JSON Schema instead of
  being buried inside an anyOf branch.
- Align ynab_create_transaction's cleared and flagColor with the YNAB enums
  (cleared/uncleared/reconciled and the six flag colors plus an empty string
  to clear a flag), replacing the boolean and unchecked cast. Apply the same
  empty-string flag value to ynab_update_transaction so both tools agree.
- Mark every tool failure with isError: true via a registry-level wrapper,
  so a failed write is never rendered as a successful call.
- Tighten date, limit, months, and transactionIds validation across several
  tools to match the constraints already used elsewhere in the repo.

* no-mistakes(review): Standardize ListBudgets failures as MCP errors

* no-mistakes(document): Correct error helper documentation

---------

+++ src/tools/CashFlowTool.ts
@@ -8,7 +8,7 @@ export const name = "ynab_cash_flow";
 export const description = "Income versus spending, month by month, so you can see whether you are running a surplus. Uses YNAB's own monthly totals rather than re-adding transactions.";
 export const inputSchema = {
   budgetId: z.string().optional().describe("The ID of the budget (optional, defaults to YNAB_BUDGET_ID environment variable)"),
-  months: z.number().positive().optional().describe("How many of the most recent months to report on (default: 6)"),
+  months: z.number().int().positive().optional().describe("How many of the most recent months to report on (default: 6)"),
   sinceDate: z.string().regex(/^\d{4}-\d{2}-\d{2}$/).optional().describe("Only include months on or after this date (ISO format: 2024-01-01). Overrides the months count."),
 };
